requirements_question4.txt
asn1crypto==0.24.0
cbpro==1.1.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coinbase==2.1.0
coinbasepro==0.3.0
crcmod==1.7
cryptography==2.6.1
distro-info==0.21
entrypoints==0.3
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.0
mysql-connector-python==8.0.25
numpy==1.20.3
protobuf==3.17.1
pycrypto==2.6.1
pycryptodome==3.10.1
pycurl==7.43.0.2
PyGObject==3.30.4
pymongo==3.5.1
python-apt==1.8.4.3
python-dateutil==2.8.1
pyxdg==0.25
requests==2.13.0
SecretStorage==2.3.1
six==1.10.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
unattended-upgrades==0.1
urllib3==1.26.5
websocket-client==0.40.0
Werkzeug==2.0.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
